refactor: workout history provider 이름변경
